Lots of refactoring (see extended notes)
Removed the text box used for input and message output

Controles are now bound to the window not a random control

Accessiblity is done "right" with a visually hidden div

Added lots of notes with tasks left to do. Everyone should read the
articles I linked in code. Will help with the report.